salt_segmentation_project/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
tqdm>=4.62.0
pyyaml>=5.4.0
albumentations>=1.0.0
scikit-image>=0.18.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
tensorboard>=2.6.0